feat: Add support for excluding specific days from tasks
This change introduces the `excludedDays` field to the `TaskRule` type, allowing users to specify days on which a particular task should not be scheduled. This improves flexibility in roster generation by accommodating recurring tasks that have weekly exceptions.

The `TaskDBModal` has been updated to allow users to toggle these excluded days for each task. A default task in `constants.ts` has also been updated to demonstrate this new functionality.